share moved to fragment
