+++ app/src/main/res/layout/activity_commodity_detail.xml
@@ -18,7 +18,7 @@
             android:layout_height="wrap_content"
             android:background="?attr/colorPrimary"
             android:minHeight="?attr/actionBarSize"
-            app:titleTextColor="@android:color/white">
+            app:titleTextColor="@color/white">
 
             <TextView
                 android:id="@+id/toolbar_title"
